refactor: ProductStatus를 domain.model 패키지로 이동
