+++ src/test/java/com/yoger/productserviceorganization/ProductServiceOrganizationApplicationTests.java
@@ -3,7 +3,7 @@ package com.yoger.productserviceorganization;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.yoger.productserviceorganization.proruct.config.AwsProperties;
-import com.yoger.productserviceorganization.proruct.domain.ProductState;
+import com.yoger.productserviceorganization.proruct.domain.model.ProductState;
 import com.yoger.productserviceorganization.proruct.dto.response.DemoProductResponseDTO;
 import java.io.IOException;
 import java.util.List;
